Add files via upload
Gradient based learning applied to digit recognition (LeNet -5, LeNet - 4, LeNet -1)

Gradient-based learning draws on the fact 
that it is generally much easier to minimize a reasonably 
smooth, continuous function than a discrete (combinatorial) 
function. The loss function can be minimized by estimating 
the impact of small variations of the parameter values on the 
loss function. This is measured by the gradient of the loss 
function with respect to the parameters. Efficient learning 
algorithms can be devised when the gradient vector can be 
computed analytically (as opposed to numerically through 
perturbations). This is the basis of numerous gradient-based 
learning algorithms with continuous-valued parameters.